cast maxDecrements

+++ script/DeployChainlinkOEVWrapper.sol
@@ -24,7 +24,7 @@ contract DeployChainlinkOEVWrapper is Script {
             addresses.getAddress("TEMPORAL_GOVERNOR"),
             addresses.getAddress("MOONWELL_WETH"),
             addresses.getAddress("WETH"),
-            uint16(10)
+            uint8(10)
         );
 
         addresses.addAddress(
